requirements.txt
python-telegram-bot==13.5
firebase-admin==5.0.0
requests==2.27.1